move dw_WriteSegmentOrExchangeData to utils.py
