+++ hydromt_delwaq/delwaq.py
@@ -18,6 +18,7 @@ from hydromt.models.model_grid import GridModel
 from tqdm import tqdm
 
 from . import DATADIR
+from .utils import dw_WriteSegmentOrExchangeData
 from .workflows import config, emissions, forcing, hydrology, segments
 
 __all__ = ["DelwaqModel"]
@@ -883,7 +884,7 @@ class DelwaqModel(GridModel):
                 fname = join(self.root, "staticdata", dvar + ".dat")
                 data = ds_out[dvar].values.flatten()
                 data = data[mask]
-                self.dw_WriteSegmentOrExchangeData(
+                dw_WriteSegmentOrExchangeData(
                     0, fname, data, 1, WriteAscii=False, mode="w"
                 )
 
@@ -1078,7 +1079,7 @@ class DelwaqModel(GridModel):
                 data = ds_out[dvar].isel(time=i).values.flatten()
                 data = data[data != nodata]
                 flowblock = np.append(flowblock, data)
-            self.dw_WriteSegmentOrExchangeData(
+            dw_WriteSegmentOrExchangeData(
                 timestepstamp[i], flname, flowblock, 1, WriteAscii=False
             )
             # volume
@@ -1090,7 +1091,7 @@ class DelwaqModel(GridModel):
                 data = ds_out[dvar].isel(time=i).values.flatten()
                 data = data[data != nodata]
                 volblock = np.append(volblock, data)
-            self.dw_WriteSegmentOrExchangeData(
+            dw_WriteSegmentOrExchangeData(
                 timestepstamp[i], voname, volblock, 1, WriteAscii=False
             )
             # sediment
@@ -1103,7 +1104,7 @@ class DelwaqModel(GridModel):
                     data = ds_out[dvar].isel(time=i).values.flatten()
                     data = data[data != nodata]
                     sedblock = np.append(sedblock, data)
-                self.dw_WriteSegmentOrExchangeData(
+                dw_WriteSegmentOrExchangeData(
                     timestepstamp[i], sedname, sedblock, 1, WriteAscii=False
                 )
             # climate
@@ -1116,7 +1117,7 @@ class DelwaqModel(GridModel):
                     data = ds_out[dvar].isel(time=i).values.flatten()
                     data = data[data != nodata]
                     climblock = np.append(climblock, data)
-                self.dw_WriteSegmentOrExchangeData(
+                dw_WriteSegmentOrExchangeData(
                     timestepstamp[i], climname, climblock, 1, WriteAscii=False
                 )
 
@@ -1315,63 +1316,6 @@ class DelwaqModel(GridModel):
             fl = fl.split(" ")
             self.set_pointer(fl, "fluxes")
         return fl
-
-    def dw_WriteSegmentOrExchangeData(
-        self,
-        ttime,
-        fname,
-        datablock,
-        boundids,
-        WriteAscii=True,
-        mode="a",
-    ):
-        """
-        Write a timestep to a segment/exchange data file.
-
-        Either appends to an existing file or creates a new one.
-
-        Input:
-            - time - timestep number for this timestep
-            - fname - File path of the segment/exchange data file</param>
-            - datablock - array with data
-            - boundids to write more than 1 block
-            - WriteAscii - if True to make a copy in an ascii checkfile
-            - mode - {"a", "w"} Force the writting mode, append or overwrite existing
-              files.
-
-        """
-        # Supress potential NaN values to avoid error (replaced by -1.0)
-        datablock[np.isnan(datablock)] = -1.0
-        # Convert the array to a 32 bit float
-        totareas = datablock
-        for i in range(boundids - 1):
-            totareas = np.vstack((totareas, datablock))
-
-        artow = np.array(totareas, dtype=np.float32).copy()
-        timear = np.array(ttime, dtype=np.int32)
-
-        if os.path.isfile(fname) and mode == "a":  # append to existing file
-            fp = open(fname, "ab")
-            tstr = timear.tobytes() + artow.tobytes()
-            fp.write(tstr)
-            if WriteAscii:
-                fpa = open(fname + ".asc", "a")
-                timear.tofile(fpa, format="%d\t", sep=":")
-                artow.tofile(fpa, format="%10.8f", sep="\t")
-                fpa.write("\n")
-        else:
-            fp = open(fname, "wb")
-            tstr = timear.tobytes() + artow.tobytes()
-            fp.write(tstr)
-            if WriteAscii:
-                fpa = open(fname + ".asc", "w")
-                timear.tofile(fpa, format="%d\t", sep=":")
-                artow.tofile(fpa, format="%10.8f", sep="\t")
-                fpa.write("\n")
-
-        fp.close()
-        if WriteAscii:
-            fpa.close()
 
     def write_monitoring(self, monpoints, monareas):
         """
